requirements.txt
certifi==2019.11.28
chardet==3.0.4
dill==0.3.1.1
idna==2.9
multiprocess==0.70.9
pathos==0.2.5
pkg-resources==0.0.0
pox==0.2.7
ppft==1.6.6.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
tqdm==4.43.0
urllib3==1.25.8
vk-api==11.7.0
